+++ enro-core/src/main/java/dev/enro/core/controller/usecase/ExecuteContainerOperationInstruction.kt
@@ -4,7 +4,6 @@ import dev.enro.core.NavigationContext
 import dev.enro.core.NavigationInstruction
 import dev.enro.core.findContainer
 import dev.enro.core.parentContainer
-import dev.enro.core.readOpenInstruction
 
 internal interface ExecuteContainerOperationInstruction {
     operator fun invoke(
@@ -29,7 +28,7 @@ internal class ExecuteContainerOperationInstructionImpl(): ExecuteContainerOpera
                 NavigationInstruction.ContainerOperation.Target.ActiveContainer -> "ActiveContainer"
                 is NavigationInstruction.ContainerOperation.Target.TargetContainer -> "TargetContainer(${instruction.target.key})"
             }
-            val contextKeyName = navigationContext.arguments.readOpenInstruction()!!.navigationKey::class.java.simpleName
+            val contextKeyName = navigationContext.instruction.navigationKey::class.java.simpleName
             "Failed to perform container instruction for $targetName in context with key $contextKeyName: Could not find valid container to perform instruction on"
         }
         instruction.operation.invoke(container)