+++ tests/application/src/androidTest/java/dev/enro/test/application/Utility.kt
@@ -19,9 +19,11 @@ fun ComposeTestRule.waitForNavigationHandle(
 ): NavigationHandle {
     var navigationHandle: NavigationHandle? = null
     waitUntil(5_000) {
-        val activity = runOnUiThread {
-            ActivityLifecycleMonitorRegistry.getInstance().getActivitiesInStage(Stage.RESUMED)
-                .singleOrNull() as? ComponentActivity
+        val activity = runOnIdle {
+            runOnUiThread {
+                ActivityLifecycleMonitorRegistry.getInstance().getActivitiesInStage(Stage.RESUMED)
+                    .singleOrNull() as? ComponentActivity
+            }
         } ?: return@waitUntil false
 
         var activeContext: NavigationContext<*>? = activity.navigationContext